# backend/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "logs-backend API",
    "version": "1.0.1",
    "description": "Authentication-first template with sessions, CSRF, rate limiting, roles, and password reset."
  },
  "servers": [
    { "url": "http://localhost:4000" }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sid"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "username_display": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "email_verified_at": { "type": "string", "format": "date-time", "nullable": true },
          "is_admin": { "type": "boolean" },
          "roles": { "type": "array", "items": { "type": "string" } },
          "created_at": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "username", "email", "is_admin"],
        "example": {
          "id": 1,
          "username": "alice",
          "username_display": "alice",
          "email": "alice@example.com",
          "email_verified_at": "2025-01-01T00:00:00.000Z",
          "is_admin": false,
          "roles": ["user"],
          "created_at": "2025-01-01T00:00:00.000Z"
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        },
        "example": { "error": "password must be 8-128 characters" }
      }
    }
  },
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "summary": "List all users (admin)",
        "tags": ["admin"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/User" }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Not authenticated" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/admin/users/{userId}/roles": {
      "put": {
        "summary": "Update user roles (admin)",
        "tags": ["admin"],
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roles": { "type": "array", "items": { "type": "string" } }
                },
                "required": ["roles"]
              },
              "example": { "roles": ["admin", "user"] }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Not authenticated" },
          "403": { "description": "Forbidden" },
          "404": { "description": "User not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/admin/audit-log": {
      "get": {
        "summary": "Audit log (admin)",
        "tags": ["admin"],
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "event", "in": "query", "schema": { "type": "string" }, "required": false },
          { "name": "user_id", "in": "query", "schema": { "type": "integer" }, "required": false },
          { "name": "since", "in": "query", "schema": { "type": "string", "format": "date-time" }, "required": false },
          { "name": "until", "in": "query", "schema": { "type": "string", "format": "date-time" }, "required": false },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": {
            "description": "Entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "entries": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "ts": { "type": "string", "format": "date-time" },
                          "event": { "type": "string" },
                          "user_id": { "type": "integer", "nullable": true },
                          "username": { "type": "string", "nullable": true },
                          "email": { "type": "string", "format": "email", "nullable": true },
                          "ip": { "type": "string", "nullable": true },
                          "meta": { "type": "object" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Not authenticated" },
          "403": { "description": "Forbidden" }
        }
      }
    },
...
  }
}